feat(components): split
