+++ src/components/Selector.tsx
@@ -2,15 +2,8 @@ import React, { useState } from 'react'
 import styled from 'styled-components'
 import { useTranslation } from 'next-i18next'
 
-import {
-  mainTheme,
-  List,
-  Box,
-  baseTheme,
-  HoverableText,
-  Button,
-} from 'danni-s-design-system'
-import { Background } from '.'
+import { List, HoverableText, Button } from 'danni-s-design-system'
+import { BackgroundKey, BACKGROUNDS, BACKGROUND_OPTIONS } from '.'
 
 import type {
   SelectorRow as SelectorRowProps,
@@ -51,72 +44,12 @@ const NavigationOptions = () => {
   ] as SelectorItemProps[]
 }
 
-const BACKGROUND = 'background'
+export const BACKGROUND = 'background'
 const SKIN = 'skin'
-const YELLOW = 'yellow'
-const GREEN = 'green'
-const RAINBOW = 'rainbow'
-const ASEXUAL = 'asexual'
-const RAINBOW_COLOUR =
-  'linear-gradient(180deg, #FE0000 16.66%,#FD8C00 16.66%, 33.32%,#FFE500 33.32%, 49.98%,#119F0B 49.98%, 66.64%,#0644B3 66.64%, 83.3%,#C22EDC 83.3%)'
-const ASEXUAL_COLOUR =
-  'linear-gradient(180deg, #181818 25%, #A3A3A3 25%, 50%, #FFFFFF 50%, 75%, #800080 75%)'
 
 const SELECTOR_NAMES = [BACKGROUND, SKIN]
 
 type SelectorName = keyof typeof SELECTOR_NAMES
-
-const BACKGROUND_OPTIONS = {
-  [GREEN]: <Background colour={mainTheme.colours.accentDark} />,
-  [YELLOW]: <Background colour="yellow" />,
-  [RAINBOW]: <Background colour={RAINBOW_COLOUR} />,
-  [ASEXUAL]: <Background colour={ASEXUAL_COLOUR} />,
-}
-
-type BackgroundKey = keyof typeof BACKGROUND_OPTIONS
-
-const BackgroundWrapper: React.FC = ({ children }) => (
-  <Box
-    height={`${baseTheme.space.xxxl}px`}
-    width={`${baseTheme.space.xxxl}px`}
-    my="m"
-    mr="m"
-    inlineBlock
-  >
-    {children}
-  </Box>
-)
-
-const BACKGROUNDS = [
-  {
-    name: BACKGROUND,
-    id: GREEN,
-    children: (
-      <BackgroundWrapper>{BACKGROUND_OPTIONS[GREEN]}</BackgroundWrapper>
-    ),
-  },
-  {
-    name: BACKGROUND,
-    id: YELLOW,
-    children: (
-      <BackgroundWrapper>{BACKGROUND_OPTIONS[YELLOW]}</BackgroundWrapper>
-    ),
-  },
-  {
-    name: BACKGROUND,
-    id: RAINBOW,
-    children: (
-      <BackgroundWrapper>{BACKGROUND_OPTIONS[RAINBOW]}</BackgroundWrapper>
-    ),
-  },
-  {
-    name: BACKGROUND,
-    id: ASEXUAL,
-    children: (
-      <BackgroundWrapper>{BACKGROUND_OPTIONS[ASEXUAL]}</BackgroundWrapper>
-    ),
-  },
-]
 
 export const Selector: React.FC<SelectorProps> = ({
   avatar,
